update postgres CreateDestination.sql
Update postgres create database statement to just check for the existence of the DB and throw a (divide by 0) exception if it does not exist. This is added as a stopgap because postgres does not support IF NOT EXISTS of "CREATE DATABASE IF NOT EXISTS mydb" without workarounds, the workarounds do not work well with transactions.